Revert "UPDATE"
This reverts commit 43249b38aec03b3279d43a645ece442ebe306642.

+++ service/user-service/src/main/java/site/toeicdoit/user/domain/vo/Messenger.java
@@ -12,12 +12,8 @@ import org.springframework.stereotype.Component;
 @AllArgsConstructor
 @NoArgsConstructor
 public class Messenger {
-    // String 메세지 전달
     private String message;
-    // Boolean 값 전달
     private Boolean state;
-    // Integer / Long 값 전달
-    private Long count;
-    // Single / List of DTO 전달
+    private Integer count;
     private Object data;
 }
